embeddings-generator/requirements.txt
python_dotenv
langchain
sentence_transformers